Arreglando issues

Closes #52, #53, #54, #55, and #56

See merge request !59

+++ src/main/java/equipo3/ujaen/backend/sistemagestioneventos/interfaces/InterfaceSistemaGestionEventos.java
@@ -13,7 +13,7 @@ public interface InterfaceSistemaGestionEventos {
 
 	public UsuarioDTO loginUsuario(String login, String password);
 
-	public List<EventoDTO> listarEventos(CategoriaEvento categoria, String descripcionParcial, long cantidadMaxima);
+	public List<EventoDTO> listarEventos(CategoriaEvento categoria, String descripcionParcial, int cantidadMaxima);
 
 	public List<EventoDTO> listarEventosInscritosDeUnUsuario(UsuarioDTO usuarioDTO);
